Fixed Overwinter decode fails (#4)
* Update transaction

Fixed variable typo.
Fixed default version from 3 to 1.

* Update transaction.js

serOverwinter is now = function and not array. This allows it to called in the application.

Reverted two lines back to their original. Needs tested with zcash and we can keep. If it doesn't work with zcash we need to set a if then most likely.

* Update transaction.js

Combined byte length from 4+4 to just 8.